pyproject.toml
[project]
name = "ai-resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain_ollama",
    "google-genai>=1.16.1",
    "browser-use==0.2.4",
    "langchain-google-genai==2.1.2",
]
